requirements.txt
pyhive
pymongo
mysql-connector-python
thrift
thrift_sasl
